requirements.txt
flask
flask_sqlalchemy
openai
python-dotenv